feature/init-setup create src folder, and move draw and mark to a separate file
